requirements.txt
flask==3.0.2
cohere==4.47
python-dotenv==1.0.1 